requirements.txt
# ------------------------------------------------------------------------------
# Python==3.11
# ------------------------------------------------------------------------------

# Core
# ------------------------------------------------------------------------------
requests==2.31.0
python-decouple==3.8
python-multipart==0.0.6

# Scraper
# ------------------------------------------------------------------------------
selenium==4.9.1
webdriver-manager==3.8.6

# Web App
# ------------------------------------------------------------------------------
fastapi==0.95.2
uvicorn==0.22.0
tortoise-orm[aiosqlite]==0.19.3
fastapi-pagination==0.12.4

# Testing
# ------------------------------------------------------------------------------
pytest==7.3.1
httpx==0.24.1

# Code quality
# ------------------------------------------------------------------------------
flake8==6.0.0  # https://github.com/PyCQA/flake8
flake8-isort==6.0.0  # https://github.com/gforcada/flake8-isort
black==23.3.0  # https://github.com/psf/black